kernel: staging/mt7621-dma: fix build conflict
While this change is not required for 5.4, it will help when we enable
kernel v5.10.

With Linux 5.10 this driver cannot be built because we have two modules
with the same name and it leads to an error:

  error: the following would cause module name conflict:
    drivers/dma/mediatek/mtk-hsdma.ko
    drivers/staging/mt7621-dma/mtk-hsdma.ko

The fix is to rename mtk-hsdma.c to hsdma-mt7621.c.  Also we can use the
KBUILD_MODNAME where appropriate instead of hard coding the name.

This patch has also been upstreamed.
Link: https://lore.kernel.org/driverdev-devel/YBky8lfQ1DmuKtRL@kroah.com/

Signed-off-by: Ilya Lipnitskiy <ilya.lipnitskiy@gmail.com>