ramips: mt7621: remove redundant lpj re-calibration
When the patch was originally introduced [0], it claimed that on mt7621,
CPU core 1 came up with a bad BogoMIPS value because the clock needed
time to stabilize. That is not the case today, at least on my Ubiquity
EdgeRouter X device. As evident from messages below, the LPJ value is
already correct after the original delay loop calibration.

Original calibration:
  [    0.033728] Calibrating delay loop... 583.68 BogoMIPS (lpj=1167360)

Redundant re-calibration (this patch removes it):
  [    0.788770] 4 CPUs re-calibrate udelay(lpj = 1167360)

So, remove the re-calibration patch to avoid extra busy looping to come
up with the exact same lpj value.

This patch also disables CONFIG_GENERIC_CLOCKEVENTS_BROADCAST as it does
not seem necessary for mt7621 - ER-X boots and functions correctly
without that symbol and no other ramips targets enable it.

This change is only for Linux 5.10, it does not affect Linux 5.4 builds.

Lastly, refresh existing patches to account for the removal of the
re-calibration logic.

[0]: 6acb53c52 ("ralink: fix rcu_sched stalls on mt7621")

Signed-off-by: Ilya Lipnitskiy <ilya.lipnitskiy@gmail.com>
Cc: John Crispin <john@phrozen.org>