ramips: drop gmac label patch, rename at preinit
Ubiquity ER-X and ER-X-SFP devices are the only ones that have a port
labeled 'eth0', which also is the default device name for the
mtk_soc_eth driver. The current solution for that involves labeling the
gmac as 'dsa' and patching mtk_soc_eth to use the label as the device
name, if available.

To avoid patching the driver, we can use 'wan' or another temporary
name for the DSA slave in device tree, then rename the gmac/phy device
at preinit to free up 'eth0', then finally rename 'wan' to 'eth0' to
match the label on the router package.

Here are the relevant boot messages with this change applied on an ER-X
device:
[    5.095343] mtk_soc_eth 1e100000.ethernet eth0: mediatek frame engine at 0xbe100000, irq 20
[    5.212708] mt7530 mdio-bus:1f wan (uninitialized): PHY [dsa-0.0:00] driver [Generic PHY] (irq=POLL)
[   10.222691] mtk_soc_eth 1e100000.ethernet dsa: renamed from eth0
[   10.278705] mt7530 mdio-bus:1f eth0: renamed from wan

Signed-off-by: Ilya Lipnitskiy <ilya.lipnitskiy@gmail.com>